refactor TreloraService for UpdateService and same folder as others
